pyproject.toml
[tool.poetry]
name = "energy-weather-pipeline"
version = "0.1.0"
description = "Weather data ingestion and transformation pipeline for ML forecasting"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Database
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
geoalchemy2 = "^0.14.3"

# Airflow
apache-airflow = "^2.8.1"
apache-airflow-providers-postgres = "^5.10.0"

# API
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"

# HTTP client
httpx = "^0.26.0"
tenacity = "^8.2.3"

# Geospatial
shapely = "^2.0.2"
geojson = "^3.1.0"

# Data quality
great-expectations = "^0.18.8"

# Utilities
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
pandas = "^2.1.4"
numpy = "^1.26.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
ipython = "^8.20.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"